requirements.txt
fastapi
uvicorn[standard]
openai
httpx
aiosqlite
python-dotenv
